pyproject.toml
[project]
name = "my-langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langchain-tavily>=0.2.13",
    "langsmith>=0.4.46",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
]
